requirements.txt
langchain==1.1.3
langchain-openai==1.1.3
streamlit==1.52.1
pypdf==6.4.1
fastapi==0.110.0
uvicorn==0.23.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.4.0
sqlalchemy==2.0.36
chonkie[genie]
